[refactor] Separate LogIn form
- setperate log in form for code readability
